Fixed a deadlock with syncing

+++ aksw-commons-io-parent/aksw-commons-io-buffers/src/main/java/org/aksw/commons/io/cache/AdvancedRangeCacheConfig.java
@@ -4,7 +4,7 @@ import java.time.Duration;
 
 public interface AdvancedRangeCacheConfig {
     int getPageSize();
-    int getMaxRequestSize();
+    long getMaxRequestSize();
     Duration getTerminationDelay();
     int getInternalWorkerSize();
 }
